+++ packages/openapi/src/decorators/property.ts
@@ -5,7 +5,7 @@ import { XSchemaObject } from '@ts-stack/openapi-spec';
 export type AnyEnum<T extends number | string = number | string> = Record<T, T>;
 
 export interface CustomType {
-  array?: Type<AnyObj> | Type<AnyObj>[];
+  array?: Type<AnyObj> | any[];
   enum?: AnyEnum | AnyEnum[];
 }
 export type PropertyDecoratorFactory = (schema?: XSchemaObject, customType?: CustomType) => PropertyDecorator;